% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_unam}
\alias{scale_colour_unam}
\alias{scale_fill_unam}
\alias{scale_color_unam}
\title{Colour scales for unam colors}
\usage{
scale_colour_unam(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_unam(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_color_unam(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in unam_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Colour scales for unam colors
}
